requirements.txt
pandas
python-telegram-bot[webhooks]
requests